// src/features/tasks/data/schema.ts
import { z } from 'zod'

// We're keeping a simple non-relational schema here.
// IRL, you will have a schema for your data models.
export const taskSchema = z.object({
  id: z.string(),
  title: z.string(),
  description: z.string().optional(),
  status: z.string(),
  label: z.string(),
  priority: z.string(),
  assignee: z.string().optional(),
  dueDate: z.union([z.string(), z.date()]).optional(),
})

export type Task = z.infer<typeof taskSchema>